clean up

+++ create_plugin_template.js
@@ -2,20 +2,20 @@
 
 /**
  * Copyright 2015 SpinalCom - www.spinalcom.com
- * 
+ *
  * This file is part of SpinalCore.
- * 
+ *
  * Please read all of the following terms and conditions
  * of the Free Software license Agreement ("Agreement")
  * carefully.
- * 
+ *
  * This Agreement is a legally binding contract between
  * the Licensee (as defined below) and SpinalCom that
  * sets forth the terms and conditions that govern your
  * use of the Program. By installing and/or using the
  * Program, you agree to abide by all the terms and
  * conditions stated or referenced herein.
- * 
+ *
  * If you do not agree to abide by these terms and
  * conditions, do not demonstrate your acceptance and do
  * not install or use the Program.
@@ -24,27 +24,27 @@
  * <http://resources.spinalcom.com/licenses.pdf>.
  */
 
-var fs = require('fs');
-var path = require('path');
+var fs = require("fs");
+var path = require("path");
 var package_path, rootPath;
 
-package_path = path.resolve('./package.json');
-rootPath = path.resolve('../..');
-modulePath = path.resolve('.');
+package_path = path.resolve("./package.json");
+rootPath = path.resolve("../..");
+let modulePath = path.resolve(".");
 
-package_path = path.resolve(modulePath + '/package.json');
-var templatePath = path.resolve(modulePath + '/templates');
+package_path = path.resolve(modulePath + "/package.json");
+var templatePath = path.resolve(modulePath + "/templates");
 
-var browserPath = path.resolve(rootPath + '/.browser_organs');
-var templatesOutputPath = path.resolve(browserPath + '/templates');
+var browserPath = path.resolve(rootPath + "/.browser_organs");
+var templatesOutputPath = path.resolve(browserPath + "/templates");
 
 function main() {
   create_folder_if_not_exit(browserPath);
   create_folder_if_not_exit(templatesOutputPath);
   if (fs.existsSync(templatePath)) {
-    var name = JSON.parse(fs.readFileSync(package_path, 'utf8')).name;
-    var src = path.resolve(templatePath + '/' + name);
-    var dst = path.resolve(templatesOutputPath + '/' + name);
+    var name = JSON.parse(fs.readFileSync(package_path, "utf8")).name;
+    var src = path.resolve(templatePath + "/" + name);
+    var dst = path.resolve(templatesOutputPath + "/" + name);
     copyRecursiveSync(src, dst);
   }
 }
@@ -68,13 +68,14 @@ function copyRecursiveSync(src, dest) {
     if (!destIsDirectory) {
       fs.mkdirSync(dest);
     }
-    fs.readdirSync(src).forEach(function (childItemName) {
-      copyRecursiveSync(path.join(src, childItemName),
-        path.join(dest, childItemName));
+    fs.readdirSync(src).forEach(function(childItemName) {
+      copyRecursiveSync(
+        path.join(src, childItemName),
+        path.join(dest, childItemName)
+      );
     });
   } else {
-    if (destExists)
-      fs.unlinkSync(dest);
+    if (destExists) fs.unlinkSync(dest);
     fs.linkSync(src, dest);
   }
 }
